+++ src/main/java/net/pincette/json/streams/Application.java
@@ -19,7 +19,7 @@ import picocli.CommandLine.HelpCommand;
     version = Application.APP_VERSION,
     description = "The JSON Streams command-line interface.")
 public class Application {
-  static final String APP_VERSION = "2.6.0";
+  static final String APP_VERSION = "2.7.0";
   private static final Logger CONFIG_LOGGER = getLogger(LOGGER_NAME + ".config");
 
   private Application() {}