+++ src/main/java/net/pincette/json/streams/Partition.java
@@ -1,17 +1,8 @@
 package net.pincette.json.streams;
 
-import static java.util.Objects.hash;
 import static java.util.Optional.ofNullable;
 
-class Partition {
-  final int offset;
-  final String topic;
-
-  Partition(final String topic, final int offset) {
-    this.topic = topic;
-    this.offset = offset;
-  }
-
+record Partition(String topic, int offset) {
   @Override
   public boolean equals(final Object o) {
     return this == o
@@ -20,10 +11,5 @@ class Partition {
             .map(Partition.class::cast)
             .filter(p -> offset == p.offset && topic.equals(p.topic))
             .isPresent();
-  }
-
-  @Override
-  public int hashCode() {
-    return hash(offset, topic);
   }
 }